0.0.52
Не большой рефактор кода
1. Для atkSpeed везде простален тип float
2. Добавлен метод валидации ValidateFloatNotNull

+++ Rouge Light Project/Assets/Scripts/Dictionary/Item/ItemImprovableCharactes.cs
@@ -6,7 +6,7 @@ public class ItemImprovableCharactes
     private string nameRu;
     private bool type;
     private int? upgradeAmount;
-    private int? finalValue;
+    private float? finalValue;
     private int? dmgUpgradeAmount;
     private int? durationUpgradeAmount;
     private int? finalDotDmg;
@@ -16,14 +16,14 @@ public class ItemImprovableCharactes
     public string Name => nameRu;       
     public bool Type => type;
     public int? UpgradeAmount => upgradeAmount;
-    public int? FinalValue => finalValue;
+    public float? FinalValue => finalValue;
     public int? DmgUpgradeAmount => dmgUpgradeAmount;  
     public int? DurationUpgradeAmount => durationUpgradeAmount;
     public int? FinalDotDmg => finalDotDmg;
     public int? FinalDotDur => finalDotDur;
 
     public ItemImprovableCharactes(string code, string nameRu, bool type, 
-        int? upgradeAmount, int? finalValue, int? dmgUpgradeAmount, 
+        int? upgradeAmount, float? finalValue, int? dmgUpgradeAmount, 
         int? durationUpgradeAmount, int? finalDotDmg, int? finalDotDur)
     {
         this.code = code;
